extern type fix
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@810 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/linux/sw_draw.c
@@ -112,7 +112,7 @@ static DWORD ScreenPtr;
 static DWORD SurfBufD;
 static DWORD *SURFDW;
 static DWORD pitch;
-extern BYTE resolutn;
+extern DWORD resolutn;
 
 void sw_clearwin()
 {
